requirements.txt
google-generativeai>=0.3.0
Pillow>=10.0.0
streamlit>=1.28.0
pathlib2>=2.3.7; python_version < "3.4"
